Revert "Init Task 8"
This reverts commit ab5aca5f6aa59c8b342bf5298361426d4b016d5d.

+++ main.h
@@ -4,14 +4,14 @@
 
 int _printf(const char *format, ...);
 int _putchar(char c);
-int print_number(long int n, const char **format, int fc);
+int print_number(int n, const char **format, int fc);
 int print_str(char *str);
 int print_binary(unsigned int n, const char **format, int fc);
 int flag_count(const char *str);
-int print_unsigned_int(long unsigned int n, const char **format, int fc);
-int print_octal(long unsigned int n, const char **format, int fc);
-int print_hex_lower(long unsigned int n, const char **format, int fc);
-int print_hex_upper(long unsigned int n, const char **format, int fc);
+int print_unsigned_int(unsigned int n, const char **format, int fc);
+int print_octal(unsigned int n, const char **format, int fc);
+int print_hex_lower(unsigned int n, const char **format, int fc);
+int print_hex_upper(unsigned int n, const char **format, int fc);
 int print_S_string(char *str);
 int print_pointer(void *p, const char **format, int fc);
